pyproject.toml
[project]
name = "meshinsights-data-pipeline"
version = "0.1.0"
description = "MeshInsights Data Pipeline"
authors = [
    { name = "Kurt Neuens", email = "kurt.neuens@meshsystems.com" }
]
requires-python = "==3.12.10"

dependencies = [
    "pyodbc==5.2.0",
    "plotly==5.24.1",
    "scipy==1.15.3",
    "scikit-learn==1.7.0",
    "diptest==0.9.0",
    "streamlit==1.45.0",
    "streamlit-authenticator==0.4.2",
    "langchain==0.3.25",
    "langgraph==0.4.2",
    "langchain-mcp-adapters==0.0.10",
    "langchain-openai==0.3.3",
    "langchain-google-genai==2.0.9",
    "langchain-anthropic==0.3.10",
    "langsmith==0.3.13",
    "openai==1.76.0",
    "openai-agents==0.0.13",
    "google-genai==1.8.0",
    "anthropic==0.49.0",
    "pandas==2.2.0",
    "numpy==1.26.4",
    "uvicorn==0.32.0",
    "python-dotenv==1.0.0",
    "pytz==2023.3",
    "pydantic==2.11.3",
    "pydantic-ai==0.1.10",
    "litellm==1.67.5",
    "pyarrow==19.0.0",
    "rapidfuzz==3.13.0",
    "mcp[cli]==1.7.0",
    "resend==2.9.0",
    "fastapi==0.115.12",
    "langserve==0.3.1",
    "matplotlib==3.10.3",
    "azure-common>=1.1.28",
    "azure-core>=1.32.0",
    "azure-identity>=1.13.0",
    "azure-iot-device>=2.12.0",
    "azure-kusto-data>=4.1.4",
    "azure-mgmt-core>=1.4.0",
    "azure-mgmt-resource>=23.0.1",
    "azure-mgmt-resourcehealth>=1.0.0b5",
    "ipykernel==6.29.5",
    "mysql-connector-python==9.4.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]          # look under /src for packages
include = ["*"]

[tool.uv]
package = true
dev-dependencies = [
    "watchdog>=2.1.0",  # Faster Streamlit reloads
    "ipython>=8.0.0",   # Better debugging
]
